update: ints to int64, floats to float64

+++ api/coins/types/id.go
@@ -7,7 +7,7 @@ type CoinID struct {
 	Symbol             string   `json:"symbol"`
 	Name               string   `json:"name"`
 	AssetPlatformID    string   `json:"asset_platform_id"`
-	BlockTimeInMinutes int      `json:"block_time_in_minutes"`
+	BlockTimeInMinutes int64    `json:"block_time_in_minutes"`
 	HashingAlgorithm   string   `json:"hashing_algorithm"`
 	Categories         []string `json:"categories"`
 	// PublicNotice ¿? `json:"public_notice"`
@@ -18,15 +18,15 @@ type CoinID struct {
 	Image                     types.Image               `json:"image"`
 	CountryOrigin             string                    `json:"country_origin"`
 	GenesisData               string                    `json:"genesis_date"`
-	SentimentVotesUpPercent   float32                   `json:"sentiment_votes_up_percentage"`
-	SentimentVotesDownPercent float32                   `json:"sentiment_votes_down_percentage"`
-	MarketCapRank             int                       `json:"market_cap_rank"`
-	CoingeckoRank             int                       `json:"coingecko_rank"`
-	CoingeckoScore            float32                   `json:"coingecko_score"`
-	DeveloperScore            float32                   `json:"developer_score"`
-	CommunityScore            float32                   `json:"community_score"`
-	LiquidityScore            float32                   `json:"liquidity_score"`
-	PublicInterestScore       float32                   `json:"public_interest_score"`
+	SentimentVotesUpPercent   float64                   `json:"sentiment_votes_up_percentage"`
+	SentimentVotesDownPercent float64                   `json:"sentiment_votes_down_percentage"`
+	MarketCapRank             int64                     `json:"market_cap_rank"`
+	CoingeckoRank             int64                     `json:"coingecko_rank"`
+	CoingeckoScore            float64                   `json:"coingecko_score"`
+	DeveloperScore            float64                   `json:"developer_score"`
+	CommunityScore            float64                   `json:"community_score"`
+	LiquidityScore            float64                   `json:"liquidity_score"`
+	PublicInterestScore       float64                   `json:"public_interest_score"`
 	MarketData                types.MarketData          `json:"market_data"`
 	CommunityData             types.CommunityData       `json:"community_data"`
 	DeveloperData             types.DeveloperData       `json:"developer_data"`
